The download URL for the SILO weather data download

Return a character string containing the parameter-formatted download URL for the SILO data, given the latitude/longitude coordinates, the start and finish dates of interest, and a list of variables to be retrieved.

Usage

build_url(latitude, longitude, station, start_date, finish_date, variables)

Arguments

latitude

The latitude (in decimal degrees North)

longitude

The longitude (in decimal degrees East)

station

The numeric BoM station ID

start_date

A string or Date object for the starting date

finish_date

A string or Date object for the finish date

variables

A vector containing the variable names

Value

The parameter-formatted URL for the data download

Examples

weathervane:::build_url(
  latitude = -34.9, longitude = 138.6,
  start = '2020-01-01', finish = '2020-12-31',
  variables = c('rainfall')
)
#> [1] "https://www.longpaddock.qld.gov.au/cgi-bin/silo/DataDrillDataset.php?lat=-34.9&lon=138.6&format=csv&username=biometryhubdev@gmail.com&password=apirequest&start=20200101&finish=20201231&comment=R"
weathervane:::build_url(
  station = 40004,
  start = '2020-01-01', finish = '2020-12-31',
  variables = c('rainfall')
)
#> [1] "https://www.longpaddock.qld.gov.au/cgi-bin/silo/PatchedPointDataset.php?station=40004&format=csv&username=biometryhubdev@gmail.com&password=apirequest&start=20200101&finish=20201231&comment=R"