Get a list of all weather stations

Get the complete list of weather stations with data available on SILO (approximately 8000).

Usage

get_all_stations(sort_by = "name")

Arguments

sort_by

The column to sort the stations by. Valid values are "name" (the default), "distance" (from Alice Springs), "id" or "state".

Value

A data.frame with all the weather stations along with their BoM station ID, Station name, Latitude, Longitude, State and Elevation.

Examples

head(get_all_stations())
#>      ID                             Name Latitude Longitude State Elevation
#> 1 41497                 AAC DALBY CAMPUS -27.1494  151.2894   QLD       347
#> 2 40000                       ABBOTSFORD -27.9500  153.1000   QLD        48
#> 3 39000                         ABERCORN -25.1333  151.1292   QLD       200
#> 4 63000 ABERCROMBIE (ABERCROMBIE BRIDGE) -33.9541  149.3252   NSW       435
#> 5 61000               ABERDEEN (MAIN RD) -32.1655  150.8911   NSW       183
#> 6 61065              ABERDEEN (ROSSGOLE) -32.1402  150.7285   NSW       543
head(get_all_stations(sort_by = "id"))
#>     ID             Name Latitude Longitude State Elevation
#> 1 1001      OOMBULGURRI -15.1806  127.8456    WA       2.0
#> 2 1005     WYNDHAM PORT -15.4644  128.1000    WA      20.0
#> 3 1006     WYNDHAM AERO -15.5100  128.1503    WA       3.8
#> 4 1009         KURI BAY -15.4875  124.5222    WA      12.0
#> 5 1010            THEDA -14.7885  126.4963    WA     210.0
#> 6 1012 MITCHELL PLATEAU -14.7925  125.8258    WA     315.0