Get list of all weather stations within specified distance

Get a list of weather stations within a provided distance from a given weather station.

Usage

get_stations_by_dist(station, distance, sort_by = "name")

Arguments

station

The station to retrieve details of. Station names will be attempted to be interpreted with an error returned.

distance

Radius from provided station.

sort_by

The column to sort the stations by. Valid values are "name" (the default), "id" or "state".

Value

A data.frame with all the weather stations along with their BoM station ID, Station name, Latitude, Longitude, State and Elevation.

Examples

get_stations_by_dist("Adelaide", 100)
#> Warning: Provided station matched multiple locations.
#>       ID                                     Name Latitude Longitude State
#> 1  14092               ADELAIDE RIVER POST OFFICE  -13.237   131.104    NT
#> 2  23000   ADELAIDE (WEST TERRACE / NGAYIRDAPIRA)  -34.926   138.583    SA
#> 3  23001                      ADELAIDE (BRIGHTON)  -35.020   138.520    SA
#> 4  23005                   ADELAIDE (GLEN OSMOND)  -34.946   138.652    SA
#> 5  23011                           NORTH ADELAIDE  -34.916   138.595    SA
#> 6  23014                      ADELAIDE (GLENSIDE)  -34.947   138.630    SA
#> 7  23016     ADELAIDE (PORT DOCK RAILWAY STATION)  -34.510   138.517    SA
#> 8  23018                ADELAIDE (TORRENS ISLAND)  -34.782   138.522    SA
#> 9  23023        ADELAIDE (SALISBURY BOWLING CLUB)  -34.767   138.643    SA
#> 10 23024                        ADELAIDE (SEATON)  -34.897   138.510    SA
#> 11 23026                       ADELAIDE (POORAKA)  -34.832   138.613    SA
#> 12 23027                 ADELAIDE (THORNDON PARK)  -34.874   138.686    SA
#> 13 23029                         ADELAIDE (UNLEY)  -34.950   138.606    SA
#> 14 23031               ADELAIDE (WAITE INSTITUTE)  -34.970   138.633    SA
#> 15 23033                ADELAIDE (TOWNSEND HOUSE)  -35.013   138.515    SA
#> 16 23034                         ADELAIDE AIRPORT  -34.952   138.520    SA
#> 17 23037 ADELAIDE (PARAFIELD PLANT INTRODUCTION C  -34.783   138.624    SA
#> 18 23040               ADELAIDE (TRINITY GARDENS)  -34.917   138.650    SA
#> 19 23057    ADELAIDE (NORTHFIELD RESEARCH CENTRE)  -34.853   138.652    SA
#> 20 23072     ADELAIDE (MAGILL SECURE CARE CENTRE)  -34.907   138.687    SA
#> 21 23075                       ADELAIDE (CLAPHAM)  -34.996   138.604    SA
#> 22 23079           ADELAIDE (DRY CREEK SALTWORKS)  -34.828   138.578    SA
#> 23 23090                     ADELAIDE (KENT TOWN)  -34.921   138.622    SA
#> 24 23096         ADELAIDE (HOPE VALLEY RESERVOIR)  -34.856   138.684    SA
#> 25 23097  ADELAIDE (BLACK HILL CONSERVATION PARK)  -34.883   138.717    SA
#> 26 23098      ADELAIDE (MORPHETTVILLE RACECOURSE)  -34.974   138.542    SA
#> 27 23732                 ADELAIDE (MORPHETT VALE)  -35.135   138.527    SA
#> 28 23748        ADELAIDE (TEA TREE GULLY COUNCIL)  -34.831   138.705    SA
#> 29 33000                            ADELAIDE PARK  -23.100   150.700   QLD
#>    Elevation
#> 1       52.5
#> 2       29.3
#> 3        8.0
#> 4      128.0
#> 5       26.0
#> 6       70.0
#> 7        3.0
#> 8        1.0
#> 9       32.0
#> 10      10.0
#> 11      21.0
#> 12      85.0
#> 13      45.7
#> 14     115.0
#> 15       5.0
#> 16       2.0
#> 17      16.0
#> 18      70.0
#> 19      77.0
#> 20     170.0
#> 21     110.0
#> 22       3.0
#> 23      48.0
#> 24     105.0
#> 25     225.0
#> 26      11.0
#> 27      92.0
#> 28     145.0
#> 29     152.4
#> Please select station ID: 
#> Error: Server-side error: Invalid start/end date