Get list of all weather stations within specified distance

Get a list of weather stations within a provided distance from a given weather station.

Usage

get_stations_by_dist(station, distance, sort_by = "distance")

Arguments

station

The station to retrieve details of. Station names will be attempted to be interpreted with an error returned.

distance

Radius in km from provided station.

sort_by

The column to sort the stations by. Valid values are "name" (the default), "distance", "id" or "state".

Value

A data.frame with all the weather stations along with their BoM station ID, Station name, Latitude, Longitude, State and Elevation.

Examples

get_stations_by_dist("Waite", 5)
#>       ID                                   Name Latitude Longitude State
#> 1  23031             ADELAIDE (WAITE INSTITUTE) -34.9697  138.6331    SA
#> 2  23010                    MITCHAM POST OFFICE -34.9833  138.6167    SA
#> 3  23014                    ADELAIDE (GLENSIDE) -34.9467  138.6303    SA
#> 4  23005                 ADELAIDE (GLEN OSMOND) -34.9464  138.6519    SA
#> 5  23029                       ADELAIDE (UNLEY) -34.9503  138.6056    SA
#> 6  23703                        BELAIR (KALYRA) -35.0042  138.6147    SA
#> 7  23075                     ADELAIDE (CLAPHAM) -34.9964  138.6039    SA
#> 8  23704           BELAIR (STATE FLORA NURSERY) -35.0076  138.6498    SA
#> 9  23746                              STONYFELL -34.9333  138.6667    SA
#> 10 23706                  BLACKWOOD POST OFFICE -35.0167  138.6167    SA
#> 11 23090                   ADELAIDE (KENT TOWN) -34.9211  138.6216    SA
#> 12 23000 ADELAIDE (WEST TERRACE / NGAYIRDAPIRA) -34.9257  138.5832    SA
#>    Elevation Distance
#> 1      115.0      0.0
#> 2       92.0      2.1
#> 3       70.0      2.3
#> 4      128.0      3.0
#> 5       45.7      3.4
#> 6      305.0      3.9
#> 7      110.0      4.0
#> 8      305.0      4.1
#> 9      250.0      5.0
#> 10     265.0      5.0
#> 11      48.0      5.0
#> 12      29.3      6.7
get_stations_by_dist(23031, 5)
#>       ID                                   Name Latitude Longitude State
#> 1  23031             ADELAIDE (WAITE INSTITUTE) -34.9697  138.6331    SA
#> 2  23010                    MITCHAM POST OFFICE -34.9833  138.6167    SA
#> 3  23014                    ADELAIDE (GLENSIDE) -34.9467  138.6303    SA
#> 4  23005                 ADELAIDE (GLEN OSMOND) -34.9464  138.6519    SA
#> 5  23029                       ADELAIDE (UNLEY) -34.9503  138.6056    SA
#> 6  23703                        BELAIR (KALYRA) -35.0042  138.6147    SA
#> 7  23075                     ADELAIDE (CLAPHAM) -34.9964  138.6039    SA
#> 8  23704           BELAIR (STATE FLORA NURSERY) -35.0076  138.6498    SA
#> 9  23746                              STONYFELL -34.9333  138.6667    SA
#> 10 23706                  BLACKWOOD POST OFFICE -35.0167  138.6167    SA
#> 11 23090                   ADELAIDE (KENT TOWN) -34.9211  138.6216    SA
#> 12 23000 ADELAIDE (WEST TERRACE / NGAYIRDAPIRA) -34.9257  138.5832    SA
#>    Elevation Distance
#> 1      115.0      0.0
#> 2       92.0      2.1
#> 3       70.0      2.3
#> 4      128.0      3.0
#> 5       45.7      3.4
#> 6      305.0      3.9
#> 7      110.0      4.0
#> 8      305.0      4.1
#> 9      250.0      5.0
#> 10     265.0      5.0
#> 11      48.0      5.0
#> 12      29.3      6.7